@@ -8,7 +8,7 @@
 class externalGravityForce
 {
 public:
-    externalGravityForce(elasticRod &m_rod, timeStepper &m_stepper, Vector3d m_gVector);
+    externalGravityForce(shared_ptr<elasticRod> m_rod, shared_ptr<timeStepper> m_stepper, Vector3d m_gVector);
     Vector3d gVector;
     void setGravity();
     ~externalGravityForce();
@@ -16,8 +16,8 @@ public:
     void computeJg();
 
 private:
-    elasticRod *rod;
-    timeStepper *stepper;
+    shared_ptr<elasticRod> rod;
+    shared_ptr<timeStepper> stepper;
     VectorXd massGravity;
 };
 
